% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-junit.R
\name{JunitReporterMeta}
\alias{JunitReporterMeta}
\title{Test reporter: summary of errors in JUnit XML format.}
\description{
This reporter includes detailed results about each test and summaries,
written to a file (or stdout) in JUnit XML format. This can be read by
the Jenkins Continuous Integration System to report on a dashboard etc.
Requires the \emph{xml2} package.
}
\details{
To fit into the JUnit structure, context() becomes the \verb{<testsuite>}
name as well as the base of the \verb{<testcase> classname}. The
test_that() name becomes the rest of the \verb{<testcase> classname}.
The deparsed expect_that() call becomes the \verb{<testcase>} name.
On failure, the message goes into the \verb{<failure>} node message
argument (first line only) and into its text content (full message).

Execution time and some other details are also recorded.

Tests without \verb{@meta} roxygen blocks produce test cases with an
empty \verb{<properties/>} node. Use \code{\link[=check_meta_test]{check_meta_test()}} to validate
that all required tags are present.

References for the JUnit XML format:
\url{http://llg.cubic.org/docs/junit/}
}
\examples{
\dontrun{
testthat::test_local(path = "inst/example.pkg", reporter = JunitReporterMeta)
}
}
\concept{reporters}
\section{Super classes}{
\code{\link[testthat:Reporter]{testthat::Reporter}} -> \code{\link[testthat:JunitReporter]{testthat::JunitReporter}} -> \code{JunitReporterMeta}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{results}}{Test results.}

\item{\code{timer}}{Process time tracker.}

\item{\code{doc}}{XML document object.}

\item{\code{errors}}{Error count for current suite.}

\item{\code{failures}}{Failure count for current suite.}

\item{\code{skipped}}{Skip count for current suite.}

\item{\code{tests}}{Test count for current suite.}

\item{\code{root}}{XML root node (\verb{<testsuites>}).}

\item{\code{suite}}{Current XML suite node (\verb{<testsuite>}).}

\item{\code{suite_time}}{Elapsed time for current suite.}

\item{\code{file_name}}{Current test file path.}

\item{\code{roxy_blocks}}{Parsed roxygen2 blocks for current file.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-JunitReporterMeta-elapsed_time}{\code{JunitReporterMeta$elapsed_time()}}
\item \href{#method-JunitReporterMeta-reset_suite}{\code{JunitReporterMeta$reset_suite()}}
\item \href{#method-JunitReporterMeta-start_reporter}{\code{JunitReporterMeta$start_reporter()}}
\item \href{#method-JunitReporterMeta-start_file}{\code{JunitReporterMeta$start_file()}}
\item \href{#method-JunitReporterMeta-start_test}{\code{JunitReporterMeta$start_test()}}
\item \href{#method-JunitReporterMeta-start_context}{\code{JunitReporterMeta$start_context()}}
\item \href{#method-JunitReporterMeta-end_context}{\code{JunitReporterMeta$end_context()}}
\item \href{#method-JunitReporterMeta-add_result}{\code{JunitReporterMeta$add_result()}}
\item \href{#method-JunitReporterMeta-end_reporter}{\code{JunitReporterMeta$end_reporter()}}
\item \href{#method-JunitReporterMeta-clone}{\code{JunitReporterMeta$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id=".start_context"><a href='../../testthat/html/Reporter.html#method-Reporter-.start_context'><code>testthat::Reporter$.start_context()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_line"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_line'><code>testthat::Reporter$cat_line()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="cat_tight"><a href='../../testthat/html/Reporter.html#method-Reporter-cat_tight'><code>testthat::Reporter$cat_tight()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_context_if_started"><a href='../../testthat/html/Reporter.html#method-Reporter-end_context_if_started'><code>testthat::Reporter$end_context_if_started()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_file"><a href='../../testthat/html/Reporter.html#method-Reporter-end_file'><code>testthat::Reporter$end_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="end_test"><a href='../../testthat/html/Reporter.html#method-Reporter-end_test'><code>testthat::Reporter$end_test()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="initialize"><a href='../../testthat/html/Reporter.html#method-Reporter-initialize'><code>testthat::Reporter$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="is_full"><a href='../../testthat/html/Reporter.html#method-Reporter-is_full'><code>testthat::Reporter$is_full()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="local_user_output"><a href='../../testthat/html/Reporter.html#method-Reporter-local_user_output'><code>testthat::Reporter$local_user_output()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="rule"><a href='../../testthat/html/Reporter.html#method-Reporter-rule'><code>testthat::Reporter$rule()</code></a></span></li>
<li><span class="pkg-link" data-pkg="testthat" data-topic="Reporter" data-id="update"><a href='../../testthat/html/Reporter.html#method-Reporter-update'><code>testthat::Reporter$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-elapsed_time"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-elapsed_time}{}}}
\subsection{Method \code{elapsed_time()}}{
Return elapsed time since last call and reset timer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$elapsed_time()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Elapsed time in seconds.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-reset_suite"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-reset_suite}{}}}
\subsection{Method \code{reset_suite()}}{
Reset suite counters to zero.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$reset_suite()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-start_reporter"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-start_reporter}{}}}
\subsection{Method \code{start_reporter()}}{
Initialize the reporter, create XML document.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$start_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-start_file"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-start_file}{}}}
\subsection{Method \code{start_file()}}{
Start processing a test file. Parses roxygen2 blocks.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$start_file(file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{Path to the test file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-start_test"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-start_test}{}}}
\subsection{Method \code{start_test()}}{
Called at the start of each test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$start_test(context, test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Test context name.}

\item{\code{test}}{Test name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-start_context"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-start_context}{}}}
\subsection{Method \code{start_context()}}{
Start a new test suite context.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$start_context(context)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Context name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-end_context"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-end_context}{}}}
\subsection{Method \code{end_context()}}{
Finalize a test suite context. Writes counts and time.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$end_context(context)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Context name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-add_result"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-add_result}{}}}
\subsection{Method \code{add_result()}}{
Add a test result. Matches \verb{@meta} tags from roxygen2
blocks and writes them as XML properties.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$add_result(context, test, result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{context}}{Context name.}

\item{\code{test}}{Test name.}

\item{\code{result}}{A testthat expectation object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-end_reporter"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-end_reporter}{}}}
\subsection{Method \code{end_reporter()}}{
Write the XML document to the output destination.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$end_reporter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-JunitReporterMeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-JunitReporterMeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JunitReporterMeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\meta{
Id: URS01.FS007

Compliance Risk: low

Implementation Complexity: medium

Test Scope: medium
}
