% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_macro_first}
\alias{move_macro_first}
\title{Move macro (meta) fragment right after arguments}
\usage{
move_macro_first(x, rd_tag = "\\\\title", offset = 1L)
}
\arguments{
\item{x}{An Rd object as returned by \code{tools::parse_Rd()}}

\item{rd_tag}{character; Rd tag name after which macro section shall be moved}

\item{offset}{integer; Offset index, so that macro block is moved after
rd_tag plus following text.}
}
\value{
transformed Rd object
}
\description{
Move macro (meta) fragment right after arguments
}
\meta{
Id: URS01.FS012

Compliance Risk: low

Implementation Complexity: low

Test Scope: low
}
