% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd_meta.R
\name{parse_rd_meta}
\alias{parse_rd_meta}
\title{Parse Metadata from Rd Files}
\usage{
parse_rd_meta(filepath)
}
\arguments{
\item{filepath}{Character array specifying paths to Rd-files.}
}
\value{
data.frame containing one column for each meta tag and one row
per filepath.
}
\description{
Parse Metadata from Rd Files
}
\meta{
Id: URS01.FS006

Compliance Risk: low

Implementation Complexity: low

Test Scope: low
}
