% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxyreqs-package.R
\docType{package}
\name{roxyreqs-package}
\alias{roxyreqs}
\alias{roxyreqs-package}
\title{roxyreqs: Roxygen2-Style Metadata for Test Cases and Function Documentation}
\description{
Extends 'roxygen2' to support '@meta' tags for documenting 'testthat' test cases and function specifications. Includes a custom JUnit reporter that exports test metadata as XML properties and validation functions to ensure all exported functions and tests contain required tags. Designed for traceability between requirements and tests in regulated industries such as pharma and finance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mnlang/roxyreqs}
  \item Report bugs at \url{https://github.com/mnlang/roxyreqs/issues}
}

}
\author{
\strong{Maintainer}: Moritz Lang \email{moritz.n.lang@gmail.com}

Authors:
\itemize{
  \item Mario Annau \email{mario.annau@gmail.com}
}

Other contributors:
\itemize{
  \item Doug Kelkhoff \email{doug.kelkhoff@gmail.com} [contributor]
  \item Szymon Maksymiuk \email{sz.maksymiuk@gmail.com} [contributor]
}

}
\keyword{internal}
